Refactor User Controller

+++ PrestigeRentals.Server/Application/Services/Interfaces/IUserManagementService.cs
@@ -13,14 +13,14 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// </summary>
         /// <param name="userId">The ID of the user to be promoted.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> MakeAdmin(int userId);
+        Task<bool> MakeAdmin(long userId);
 
         /// <summary>
         /// Reverts a user from an admin role to a regular user.
         /// </summary>
         /// <param name="userId">The ID of the user to be reverted.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> RevertToUser(int userId);
+        Task<bool> RevertToUser(long userId);
 
         /// <summary>
         /// Changes the password for a user.
@@ -29,7 +29,7 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// <param name="oldPassword">The user's current password.</param>
         /// <param name="newPassword">The new password to be set for the user.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> ChangePassword(int userId, string oldPassword, string newPassword);
+        Task<bool> ChangePassword(long userId, string oldPassword, string newPassword);
 
         /// <summary>
         /// Changes the email address for a user.
@@ -37,28 +37,28 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// <param name="userId">The ID of the user whose email will be changed.</param>
         /// <param name="newEmail">The new email address to be set for the user.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> ChangeEmail(int userId, string newEmail);
+        Task<bool> ChangeEmail(long userId, string newEmail);
 
         /// <summary>
         /// Deactivates the account of a user.
         /// </summary>
         /// <param name="userId">The ID of the user whose account will be deactivated.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> DeactivateAccount(int userId);
+        Task<bool> DeactivateAccount(long userId);
 
         /// <summary>
         /// Activates the account of a deactivated user.
         /// </summary>
         /// <param name="userId">The ID of the user whose account will be activated.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> ActivateAccount(int userId);
+        Task<bool> ActivateAccount(long userId);
 
         /// <summary>
         /// Deletes the account of a user.
         /// </summary>
         /// <param name="userId">The ID of the user whose account will be deleted.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> DeleteAccount(int userId);
+        Task<bool> DeleteAccount(long userId);
 
         /// <summary>
         /// Updates the details of a user (e.g., first name, last name).
@@ -66,6 +66,6 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// <param name="userId">The ID of the user whose details will be updated.</param>
         /// <param name="updateUserDetailsRequest">The request containing the new details for the user.</param>
         /// <returns>A task representing the asynchronous operation, with a boolean result indicating success.</returns>
-        Task<bool> UpdateUserDetails(int userId, UpdateUserDetailsRequest updateUserDetailsRequest);
+        Task<bool> UpdateUserDetails(long userId, UpdateUserDetailsRequest updateUserDetailsRequest);
     }
 }
